backend/requirements.txt
# PropVision.AI Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
shapely==2.0.6
alembic==1.14.1
redis==5.2.1
httpx==0.28.1
openai==1.59.3
python-multipart==0.0.20
boto3==1.36.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
email-validator==2.2.0
apscheduler==3.10.4
tenacity==9.0.0
orjson==3.10.14

